Minor updates
Added a cursor lock check for when the player is in the game, which locks it to the center of the game window, but when pausing, the lock is removed. Also made sure player can't move around during pause menu screen.

Originally a player was only able to host one game and after leaving could not host another game as the NetworkManager object which followed into our game session through scene changes was not attached back to the lobby scene, causing the on click event of hosting a game to be disabled and not allow a player to create another room. This has now been fixed and a player can create multiple rooms.

The last update was to the Join Game room list which appears in our lobby scene. A player could host a room and another player could see it in their join game list, but if the room host left, the other player would seemingly try to join indefinitely. I updated the code to use a Enumerator  which would take a countdown variable for players joining a room, and if they could not join within the given amount of time, they would be prompted with a "failed to join" message, which would then update the join game room list to properly mimick if the host game has been dropped and is unavailable to be joined again.